voting_docker_service/requirements.txt
astunparse==1.6.3
brotli==1.1.0
cffi==1.17.1
cmake==4.0.3
colorama==0.4.6
conda-build==25.7.0
conda-index==0.6.1
conda-libmamba-solver==25.3.0
exceptiongroup==1.3.0
expecttest==0.3.0
fastapi==0.133.1
h2==4.2.0
hdbscan==0.8.41
httpx==0.28.1
hypothesis==6.137.1
importlib-metadata==8.0.0
ipython==9.4.0
jaraco.collections==5.1.0
lief==0.16.6
lintrunner==0.12.7
ninja==1.11.1.4
optree==0.17.0
pickleshare==0.7.5
pip-chill==1.0.3
pysocks==1.7.1
tomli==2.0.1
torchaudio==2.8.0+cu128
torchelastic==0.2.2
torchvision==0.23.0+cu128
transformers==4.57.5
typer-slim==0.24.0
types-dataclasses==0.6.6
uvicorn==0.41.0
